# pongChain/backend-blockchain/openapi.yaml
openapi: 3.0.3
info:
  title: PongChain API
  description: API REST pour interagir avec le MasterContract
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /match/{id}:
    get:
      summary: Détails d’un match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Informations du match
  /tournament/{id}:
    get:
      summary: Détails d’un tournoi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Informations du tournoi
  /tournament/winner/{address}:
    get:
      summary: Tournois gagnés par une adresse
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Liste des tournois gagnés
...